requirements.txt
# A list of packages that installed by default in the pytorch/pytorch docker image
# torch==1.13.0
# torchtext==0.14.0
# torchvision==0.14.0

# Transformer Lens V2
transformer-lens @ git+https://github.com/neelnanda-io/TransformerLens@25b9e076fd6a44d053f5e2925a5d1ff802692ab2
protobuf==3.20

# Minigrid
Minigrid==2.1.0

# torch optimizer (for lamb)
torch-optimizer==0.3.0

# tqdm
tqdm==4.64.1

# kaleido for figures
kaleido==0.2.1

# Jupyter for notebook dev
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4

# plotting
plotly==5.10.0
matplotlib==3.6.2

# pytest
pytest==7.1.3
pytest-cov==4.0.0

# wandb
wandb==0.13.7

#numpy
numpy==1.23.5

#moviepy
moviepy==1.0.3

# streamlit
streamlit==1.16.0
